# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intervalo Backend",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/course/{course_id}/belts": {
      "get": {
        "summary": "Get Belt Info",
        "description": "Returns descriptive info (headline + description) for each belt in a course.",
        "operationId": "get_belt_info_course__course_id__belts_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/BeltEntry"
                  },
                  "title": "Response Get Belt Info Course  Course Id  Belts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get Current User Info",
        "description": "Get current authenticated user info.",
        "operationId": "get_current_user_info_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/enroll": {
      "post": {
        "summary": "Enroll User",
        "description": "Enroll user in a course with onboarding data.",
        "operationId": "enroll_user_user_enroll_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/progress": {
      "get": {
        "summary": "Get User Progress",
        "description": "Get user's current progress (topic states and level info).",
        "operationId": "get_user_progress_user_progress_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/start": {
      "post": {
        "summary": "Start Session",
        "description": "Start a new session linked to authenticated user in database.",
        "operationId": "start_session_session_start_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/start-zen": {
      "post": {
        "summary": "Start Zen Session",
        "description": "Start a Zen session: random exercises from selected belts, no SM-2 logic.",
        "operationId": "start_zen_session_session_start_zen_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartZenSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/answer": {
      "post": {
        "summary": "Submit Answer",
        "description": "Submit answer for exercise - validates ownership and saves to DB.",
        "operationId": "submit_answer_session_answer_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/session/{session_id}/summary": {
      "get": {
        "summary": "Session Summary",
        "description": "Get session summary from database - validates ownership.",
        "operationId": "session_summary_session__session_id__summary_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "exercise_id": {
            "type": "string",
            "title": "Exercise Id"
          },
          "answer_index": {
            "type": "integer",
            "title": "Answer Index"
          },
          "response_time_s": {
            "type": "number",
            "title": "Response Time S"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "exercise_id",
          "answer_index",
          "response_time_s"
        ],
        "title": "AnswerRequest"
      },
      "BeltEntry": {
        "properties": {
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "headline",
          "description"
        ],
        "title": "BeltEntry"
      },
      "BeltProgressInfo": {
        "properties": {
          "mastered": {
            "type": "integer",
            "title": "Mastered"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "promoted": {
            "type": "boolean",
            "title": "Promoted"
          }
        },
        "type": "object",
        "required": [
          "mastered",
          "total",
          "promoted"
        ],
        "title": "BeltProgressInfo"
      },
      "EnrollmentRequest": {
        "properties": {
          "university": {
            "type": "string",
            "title": "University"
          },
          "career": {
            "type": "string",
            "title": "Career"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "university",
          "career"
        ],
        "title": "EnrollmentRequest"
      },
      "EnrollmentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "EnrollmentResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "LevelInfo": {
        "properties": {
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "xp_in_level": {
            "type": "integer",
            "title": "Xp In Level"
          },
          "xp_required": {
            "type": "integer",
            "title": "Xp Required"
          },
          "progress_pct": {
            "type": "number",
            "title": "Progress Pct"
          }
        },
        "type": "object",
        "required": [
          "level",
          "xp_in_level",
          "xp_required",
          "progress_pct"
        ],
        "title": "LevelInfo"
      },
      "LevelInfoWithMissing": {
        "properties": {
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "xp_in_level": {
            "type": "integer",
            "title": "Xp In Level"
          },
          "xp_required": {
            "type": "integer",
            "title": "Xp Required"
          },
          "xp_missing": {
            "type": "integer",
            "title": "Xp Missing"
          },
          "progress_pct": {
            "type": "number",
            "title": "Progress Pct"
          }
        },
        "type": "object",
        "required": [
          "level",
          "xp_in_level",
          "xp_required",
          "xp_missing",
          "progress_pct"
        ],
        "title": "LevelInfoWithMissing"
      },
      "SessionExercise": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "correct_index": {
            "type": "integer",
            "title": "Correct Index"
          },
          "has_math": {
            "type": "boolean",
            "title": "Has Math"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "belt": {
            "type": "string",
            "title": "Belt"
          },
          "graph_fn": {
            "type": "string",
            "title": "Graph Fn"
          },
          "graph_view": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph View"
          },
          "feedback_correct": {
            "type": "string",
            "title": "Feedback Correct"
          },
          "feedback_incorrect": {
            "type": "string",
            "title": "Feedback Incorrect"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question",
          "options",
          "correct_index",
          "has_math",
          "topic",
          "belt",
          "graph_fn",
          "feedback_correct",
          "feedback_incorrect"
        ],
        "title": "SessionExercise"
      },
      "SessionStartResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/SessionExercise"
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_name",
          "total",
          "exercises"
        ],
        "title": "SessionStartResponse"
      },
      "SessionSummaryResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "incorrect": {
            "type": "integer",
            "title": "Incorrect"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SummaryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "topic_states": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TopicProgress"
            },
            "type": "object",
            "title": "Topic States"
          },
          "belt_progress": {
            "$ref": "#/components/schemas/BeltProgressInfo"
          },
          "xp_earned": {
            "type": "integer",
            "title": "Xp Earned"
          },
          "level_info": {
            "$ref": "#/components/schemas/LevelInfoWithMissing"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_name",
          "total",
          "correct",
          "incorrect",
          "items",
          "topic_states",
          "belt_progress",
          "xp_earned",
          "level_info"
        ],
        "title": "SessionSummaryResponse"
      },
      "StartSessionRequest": {
        "properties": {
          "user_name": {
            "type": "string",
            "title": "User Name"
          }
        },
        "type": "object",
        "required": [
          "user_name"
        ],
        "title": "StartSessionRequest"
      },
      "StartZenSessionRequest": {
        "properties": {
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "belts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Belts"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "user_name",
          "belts",
          "count"
        ],
        "title": "StartZenSessionRequest"
      },
      "SummaryItem": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "belt": {
            "type": "string",
            "title": "Belt"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "belt",
          "correct"
        ],
        "title": "SummaryItem"
      },
      "TopicProgress": {
        "properties": {
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "step_index": {
            "type": "integer",
            "title": "Step Index"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "string",
            "title": "Progress"
          },
          "is_pending": {
            "type": "boolean",
            "title": "Is Pending"
          },
          "attempted": {
            "type": "boolean",
            "title": "Attempted"
          },
          "next_review": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review"
          },
          "failed": {
            "type": "boolean",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "step_index",
          "status",
          "progress",
          "is_pending",
          "attempted",
          "failed"
        ],
        "title": "TopicProgress"
      },
      "UserProgressResponse": {
        "properties": {
          "topic_states": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TopicProgress"
            },
            "type": "object",
            "title": "Topic States"
          },
          "level_info": {
            "$ref": "#/components/schemas/LevelInfo"
          }
        },
        "type": "object",
        "required": [
          "topic_states",
          "level_info"
        ],
        "title": "UserProgressResponse"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "clerk_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clerk User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}